+++ bc-react/src/actions/MessageActions.js
@@ -1,5 +1,12 @@
 import dispatcher from '../Dispatcher'
 
+let baseUrl;
+if(process.env.NODE_ENV === 'production'){
+  baseUrl = "/"
+} else {
+  baseUrl = "http://localhost:4000/"
+}
+
 export function updateMessageInput(attribute, value){
   dispatcher.dispatch({
     type: 'UPDATE_MESSAGE_INPUT',
@@ -8,31 +15,45 @@ export function updateMessageInput(attribute, value){
   })
 }
 
-export function submitMessageInput(messageAttributes){
+export function sendMessage(messageAttributes){
   dispatcher.dispatch({
-    type: 'SUBMIT_MESSAGE_INPUT'
+    type: 'SEND_MESSAGE'
   })
 }
 
-//How could this be renamed for clarity? ie. updateMessageStore? (not sure if that's accurate)
-export function updateMessageDetail(attributes){
-  dispatcher.dispatch({
-    type: 'UPDATE_MESSAGE_DETAIL',
-    //the attributes will be just the attributes that were passed
-    attributes: attributes
+export function processMessage(attributes){
+  const params = {
+    //this is the action we send over
+    method: 'POST',
+    headers: {'Content-Type': 'application/json'},
+    body: JSON.stringify(attributes)
+  }
+  //fetch is a new feature of ES6, it's a standard for fetching data asyncrhonously using promises
+  fetch(`${baseUrl}messages`, params).then((response)=>{
+    if(response.ok){0
+      response.json().then((body)=>{
+        //this is the message we get back in postman
+        //here we handle it back on the server
+        //this returns the message that was just created
+        dispatcher.dispatch({
+          type: 'UPDATE_MESSAGE_DETAIL',
+          attributes: body.message
+        })
+      })
+    }else{
+      response.json().then((body)=>{
+        dispatcher.dispatch({
+          type: 'MESSAGE_FAIL',
+          errors: body.errors
+        })
+      })
+    }
   })
 }
 
-export function messageFail(errors){
-  dispatcher.dispatch({
-    type: 'MESSAGE_FAIL',
-    errors: errors
-  })
-}
-
-export function updateMessages(attributes){
-  dispatcher.dispatch({
-    type: 'UPDATE_MESSAGES',
-    attributes: attributes
-  })
-}
+// export function updateMessages(attributes){
+//   dispatcher.dispatch({
+//     type: 'UPDATE_MESSAGES',
+//     attributes: attributes
+//   })
+// }
